Use standard Array.sort and do not update search_count when only ids
The current implementation of Group.sort is not correct because it does not
guarantee to replace all the records if the ordered list does not contain all
the existing records. So it is better to use the standard Array.sort method.
Also the search_count should not be updated when doing a search only for ids
as the number of record does not change.

issue8042
review50731002