Reduce number of promises when rendering tree
We preload the first field of the first row if needed. This should ensure that
the following records are also loaded.
And we do not call load if the field is already loaded to avoid to create a
promise.

issue8199
review265411002
(grafted from 02a9082b44d19cc80a8af111e3c1611f0fd11593)